rag-mcp/requirements.txt
python-dotenv
requests
fastapi
uvicorn
pydantic
fastmcp